Change to match prototype (size_t vs unsigned int) so it works on 64-bit
	* camel-stream-mem.c (camel_stream_mem_new_with_buffer): Change
	to match prototype (size_t vs unsigned int) so it works on 64-bit
	machines. Noted by msw.

svn path=/trunk/; revision=2998

+++ camel/camel-stream-mem.c
@@ -109,7 +109,7 @@ camel_stream_mem_new (void)
 }
 
 CamelStream *
-camel_stream_mem_new_with_buffer (const char *buffer, unsigned int len)
+camel_stream_mem_new_with_buffer (const char *buffer, size_t len)
 {
 	GByteArray *ba;
